requirements.txt
streamlit
gspread
oauth2client
pillow
feedparser
requests
python-dotenv
openai
cloudinary